Migrate the VS import functionality to the VCS integration enabler (#67)
