+++ plugin/resources/com/microsoft/alm/plugin/idea/ui/tfplugin.properties
@@ -208,8 +208,6 @@ Tfvc.Lock.Dialog.UnlockButton=Unlock
 
 # TFVC Import Workspace Action
 Actions.Tfvc.ImportWorkspace.Title=Import TFVC Workspace
-Actions.Tfvc.ImportWorkspace.Description=Import an externally-created workspace, e.g. a Visual Studio one
-Actions.Tfvc.SelectWorkspacePath=Select Workspace Path
 Tfvc.WorkspaceNotDetected=Workspace could not be determined from the current directory. Make sure you have the right directory selected and the right Visual Studio TF client configured.
 
 #TFVC branch action
